refactor(trading-signals): share Ehlers' dominant-cycle measurement between MAMA and HT Trendline
The Hilbert transform, homodyne discriminator, and period estimation were
duplicated verbatim (~120 lines) across both indicators — the largest
clone in the codebase. The engine now lives in one place; its working
memory embeds flat into each indicator's own state so replace/rollback
snapshots keep covering the whole measurement. Emissions stay bit-for-bit
identical to the TA-Lib-anchored expectations.
